+++ frontend/src/data/archive-api.js
@@ -1,9 +1,9 @@
 import { throwIfNotSuccess, headers } from './helper';
 
-const URI = 'api/archive';
+const URI = 'api/admin/archive';
 
 function fetchAll () {
-    return fetch(`${URI}/features`)
+    return fetch(`${URI}/features`, { credentials: 'include' })
         .then(throwIfNotSuccess)
         .then(response => response.json());
 }