+++ frontend/src/data/history-api.js
@@ -1,15 +1,15 @@
 import { throwIfNotSuccess } from './helper';
 
-const URI = 'api/events';
+const URI = 'api/admin/events';
 
 function fetchAll () {
-    return fetch(URI)
+    return fetch(URI, { credentials: 'include' })
         .then(throwIfNotSuccess)
         .then(response => response.json());
 }
 
 function fetchHistoryForToggle (toggleName) {
-    return fetch(`${URI}/${toggleName}`)
+    return fetch(`${URI}/${toggleName}`, { credentials: 'include' })
         .then(throwIfNotSuccess)
         .then(response => response.json());
 }